Обработка исключений для PersonTransactionsController

+++ src/main/java/ru/ServerRestApp/controllers/PersonTransactionsController.java
@@ -5,9 +5,12 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.validation.BindingResult;
-import org.springframework.validation.FieldError;
 import org.springframework.web.bind.annotation.*;
+import ru.ServerRestApp.models.Category;
+import ru.ServerRestApp.models.CategoryTransaction;
+import ru.ServerRestApp.models.Person;
 import ru.ServerRestApp.models.PersonTransaction;
+import ru.ServerRestApp.services.PeopleService;
 import ru.ServerRestApp.services.PersonTransactionsService;
 import ru.ServerRestApp.util.ErrorResponse;
 import ru.ServerRestApp.util.DataException;
@@ -23,48 +26,107 @@ import static ru.ServerRestApp.util.ErrorsUtil.returnDataErrorsToClient;
 public class PersonTransactionsController {
 
     private final PersonTransactionsService personTransactionsService;
+    private final PeopleService peopleService;
     @Autowired
-    public PersonTransactionsController(PersonTransactionsService personTransactionsService) {
+    public PersonTransactionsController(PersonTransactionsService personTransactionsService, PeopleService peopleService) {
         this.personTransactionsService = personTransactionsService;
+        this.peopleService = peopleService;
     }
 
 
     @GetMapping()
-    public List<PersonTransaction> getAllPersonTransactions() {
-        return personTransactionsService.findAll();
+    public ResponseEntity<List<PersonTransaction>> getAllPersonTransactions() {
+        List<PersonTransaction> personTransaction = personTransactionsService.findAll();
+        return new ResponseEntity<>(personTransaction, HttpStatus.OK);
     }
 
     @GetMapping("/{id}")
-    public PersonTransaction getPersonTransaction(@PathVariable("id") int id) {
+    public ResponseEntity<PersonTransaction> getPersonTransaction(@PathVariable("id") int id) {
         Optional<PersonTransaction> personTransaction = personTransactionsService.findById(id);
-        if (personTransaction.isPresent())
-            return personTransaction.get();
-        else
+        if (personTransaction.isEmpty())
             throw new NotFoundException("PersonTransaction with this id wasn't found!");
+        return new ResponseEntity<>(personTransaction.get(), HttpStatus.OK);
     }
 
     @PostMapping("/add")
     public ResponseEntity<PersonTransaction> addPersonTransaction(@RequestBody @Valid PersonTransaction personTransaction, BindingResult bindingResult) {
 
+        personTransaction.setId(0);
         if (bindingResult.hasErrors())
             returnDataErrorsToClient(bindingResult);
+
+        if (personTransaction.getPerson_from() == null)
+            throw new DataException("Person_from must not be null!");
+
+        Optional<Person> person_from = peopleService.findById(personTransaction.getPerson_from().getId());
+        if (person_from.isEmpty())
+            throw new NotFoundException("Person_from with this id wasn't found!");
+
+        if (personTransaction.getPerson_to() == null)
+            throw new DataException("Person_to must not be null!");
+
+        Optional<Person> person_to = peopleService.findById(personTransaction.getPerson_to().getId());
+        if (person_to.isEmpty())
+            throw new NotFoundException("Person_to with this id wasn't found!");
+
+        if (personTransaction.getPerson_from().getId() == personTransaction.getPerson_to().getId())
+            throw new DataException("Person_from id must not be equal to Person_to id!");
+
+        if (personTransaction.getCreated_at() == null)
+            throw new DataException("Created_at must not be null!");
 
         personTransactionsService.save(personTransaction);
 
         return new ResponseEntity<>(personTransaction, HttpStatus.OK);
     }
 
-    @PostMapping("/update")
-    public PersonTransaction updatePersonTransaction(@RequestBody PersonTransaction personTransaction) {
+    @PostMapping("/update/{id}")
+    public ResponseEntity<PersonTransaction> updatePersonTransaction(@PathVariable("id") int id, @RequestBody @Valid PersonTransaction personTransaction, BindingResult bindingResult) {
+
+        personTransaction.setId(id);
+        if (bindingResult.hasErrors())
+            returnDataErrorsToClient(bindingResult);
+
+        Optional<PersonTransaction> foundPersonTransaction = personTransactionsService.findById(id);
+        if (foundPersonTransaction.isEmpty())
+            throw new NotFoundException("PersonTransaction with this id wasn't found!");
+
+        if (personTransaction.getPerson_from() == null)
+            throw new DataException("Person_from must not be null!");
+
+        Optional<Person> person_from = peopleService.findById(personTransaction.getPerson_from().getId());
+        if (person_from.isEmpty())
+            throw new NotFoundException("Person_from with this id wasn't found!");
+
+        if (personTransaction.getPerson_to() == null)
+            throw new DataException("Person_to must not be null!");
+
+        Optional<Person> person_to = peopleService.findById(personTransaction.getPerson_to().getId());
+        if (person_to.isEmpty())
+            throw new NotFoundException("Person_to with this id wasn't found!");
+
+        if (personTransaction.getPerson_from().getId() == personTransaction.getPerson_to().getId())
+            throw new DataException("Person_from id must not be equal to Person_to id!");
+
+        if (personTransaction.getCreated_at() == null)
+            throw new DataException("Created_at must not be null!");
+
         personTransactionsService.update(personTransaction);
-        return personTransactionsService.findById(personTransaction.getId()).get();
+
+        return new ResponseEntity<>(personTransaction, HttpStatus.OK);
     }
 
     @PostMapping("/delete/{id}")
-    public void deletePersonTransaction(@PathVariable("id") int id) {
-        personTransactionsService.delete(id);
-    }
+    public ResponseEntity<PersonTransaction> deletePersonTransaction(@PathVariable("id") int id) {
 
+        Optional<PersonTransaction> foundPersonTransaction = personTransactionsService.findById(id);
+        if (foundPersonTransaction.isEmpty())
+            throw new NotFoundException("PersonTransaction with this id wasn't found!");
+
+        personTransactionsService.delete(id);
+
+        return new ResponseEntity<>(foundPersonTransaction.get(), HttpStatus.OK);
+    }
 
 
     @ExceptionHandler
